Make container friendly
The original sample appears to have been designed only to run via the
Liberty Maven plugin and certainly not in a container. Moved
resources/CustomConfigSource.json under the Liberty server directory and
modified CustomConfigSource.java to look for it relative to their.
MetricsOverHTTPResource.java is hard-coded to use port 9443 which may
not resolve using the external URL if ports have been mapped. Switched
to localhost which should work in both the Maven and container
scenarios. Finally, added a Dockerfile as I've yet to contrive something
sufficiently generic to go in the Draft pack that would also work for
this application (requires a server.xml to set the required features,
adminSecurity and the context root, which, in turn, needs the WAR
renaming to match, and also needs the config json adding to the image).